refactor.

+++ install.sh
@@ -21,19 +21,23 @@ install_plugin $zsh_plugins_dir "zsh-users" "zsh-autosuggestions"
 install_plugin $zsh_plugins_dir "zsh-users" "zsh-syntax-highlighting"
 install_plugin $tmux_plugins_dir "tmux-plugins" "tpm"
 
-# params: source_dir, target_dir
+# params: source, target
 function link_file() {
-    if [ ! -L $2 ]; then
+    # check symlink and dir exist
+    if [[ ! -L $2 && ! -e $2 ]]; then
         echo "Linking $1 to $2"
         ln -s $1 $2
     else
-        echo "Symlink already exists: $2"
+        echo "File or directory already exists(source: $1, target: $2)"
     fi
 }
 
-link_file ~/dotfiles/.zshrc ~/.zshrc
-link_file ~/dotfiles/.gitconfig ~/.gitconfig
-link_file ~/dotfiles/.profile ~/.profile
-link_file ~/dotfiles/.tmux.conf ~/.tmux.conf
-link_file ~/dotfiles/.bashrc ~/.bashrc
-link_file ~/dotfiles/nvim ~/.config/nvim
+link_file $HOME/dotfiles/.zshrc ~/.zshrc
+link_file $HOME/dotfiles/.gitconfig ~/.gitconfig
+link_file $HOME/dotfiles/.profile ~/.profile
+link_file $HOME/dotfiles/.tmux.conf ~/.tmux.conf
+link_file $HOME/dotfiles/.bashrc ~/.bashrc
+link_file $HOME/dotfiles/nvim $HOME/.config/
+
+# copy fonts
+cp -a $HOME/dotfiles/fonts/. $HOME/.local/share/fonts/
